каша какая-то с селекторами
