+++ src/components/app/app.jsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import {connect} from 'react-redux';
-import {ActionCreator} from "../../reducer.js";
+import reducer from '../../reducer';
+import {getAllOffers, getCurrentCity, getCities} from "../../reducer/data/selectors";
 import {PageScreen} from '../page-screen/page-screen.jsx';
 
 const App = (props) => {
@@ -12,18 +13,18 @@ const App = (props) => {
 };
 
 const mapStateToProps = (state, ownProps) => Object.assign({}, ownProps, {
-  allOffers: state.allOffers,
+  allOffers: getAllOffers(state),
   isOffersLoading: state.isOffersLoading,
-  currentCity: state.currentCity,
+  currentCity: getCurrentCity(state),
   offers: state.offers,
-  cities: state.cities,
+  cities: state.cities, // getCities(state),
   isAuthorizationRequired: state.isAuthorizationRequired,
 });
 
 const mapDispatchToProps = (dispatch) => ({
   onCityClick: (currentCity, allOffers) => {
-    dispatch(ActionCreator.changeCity(currentCity));
-    dispatch(ActionCreator.getOffersList(currentCity, allOffers));
+    dispatch(reducer.ActionCreator.changeCity(currentCity));
+    dispatch(reducer.ActionCreator.getOffersList(currentCity, allOffers));
   }
 
 });
