+++ src/reducer/name-spaces.js
@@ -1,6 +1,5 @@
 const NameSpace = {
   DATA: `DATA`,
-  GAME: `WEB_APP`,
   USER: `USER`
 };
